requirements.txt
Django>=3.1.3,<3.2.0
djangorestframework>=3.12.2,<3.13.0
gunicorn>=20.0.4,<20.1.0
drf-yasg>=1.20.0,<1.21.0
flake8>=3.8.4,<3.9.0
django-cors-headers>=3.5.0,<3.6.0
python-dotenv>=0.15.0,<0.16.0